requirements.txt
-e .
aws_cdk.aws_s3
aws_cdk.aws_sqs
aws_cdk.aws_kms
aws_cdk.aws_s3_notifications
aws_cdk.aws_iam
